+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Support asking the AMS router for an open port.  This is required
   when connecting to a PLC on 127.0.0.1.
+- Support for directly reading Rust types from `Device`s and
+  `symbol::Handle`s using `read_value`/`write_value`.
 - Add more known ADS states.
 - Document the `symbol::Symbol` and `symbol::Type` flags.
 - Add an adstool command to list AMS routes on the target.